practica finalizada

+++ Aplicaciones/Academico/models.py
@@ -1,11 +1,12 @@
 from django.db import models
+from uuid import uuid4
 
 # Create your models here.
 
 class Curso(models.Model):
     codigo=models.CharField(primary_key=True, max_length=6)
     nombre=models.CharField(max_length=50)
-    creditos=models.PositiveBigIntegerField()
+    creditos=models.PositiveIntegerField()
 
     def __str__(self):
         texto = "{0} ({1})"
